requirements.txt
pycryptodome
lxml
Pillow
aiohttp